Move pass mode handling to abi/pass_mode.rs
